Committed by Max Milbers integration of multivendor.
Topic "Admin":     
The Userlist shows which user is vendor (Possibility to expand it to activating and deactivating a vendor)

Topic "Store":   
Every vendor has its own store information. If an admin wants to see this data of other vendors he need to look in the topic "Vendor".

Topic "Products":
"List Products" shows the only the products of the logged in vendor exzept he is an admin.
TODO: A new column which shows the owner of a product

"Add/Edit Product"
The admin can change the owner(vendor) of a product.
If a normal vendor is logged in, he is automatically set as vendor. He can only use his or shared categories (see categories for mor information)

"View Inventory"
As in list products,... it lists only the products that belong to the vendor, except the admin.

"Special Products"
Seems to work, vendors see only their products, except the admin

"Customer Reviews"
The vendors see only reviews of there own products, except the admin.

Categories:
The categories are full working. Some people want that a vendor can make his own categories. Other people want that all vendors use the categories together. So I made the categories shareable. it is all working very nice.


git-svn-id: https://dev.virtuemart.net/svn/virtuemart@1532 87cc427b-4a1e-0410-b729-b2260ec0f063